fix: return ffmpeg error logs to caller, and fix StreamingStatus (#254)
* fix: return ffmpeg error logs to caller, and fix StreamingStatus

FFmpeg transcoder logic has been implemented within the driver code now,
so that way we have more control over it. This allows us to keep track
of the process output and log it to the console as well as return it to
the caller of StartStreaming. Because we have more precise control over
the process, we can better sync the StreamingStatus to whether or not
the ffmpeg process is running.

The handling of publishing StreamingStatus to the message bus has been
optimized to send better based on when it actually changes, and not send
when the user calls StopStreaming when already stopped, or
StartStreaming when already started.

Logic has also been added to redact the rtsp credentials from the log
file and API responses to avoid exposing secrets to the caller.

Signed-off-by: Anthony Casagrande <anthony.j.casagrande@intel.com>